Add rr structure

+++ firmware/src/system/task/task_func.c
@@ -11,6 +11,6 @@ void task1 (void) {
 void task2 (void) {
     for (;;) {
         GPIOA->BSRR = GPIO_BSRR_BR5;
-        for(uint16_t i = 0; i < 700000; i++) {}
+        for(uint32_t i = 0; i < 700000; i++) {}
     }
 }
